Cleanup TS code (and fix rescript build warning)
