+++ packages/squiggle-lang/__tests__/TS/JS_test.ts
@@ -1,4 +1,4 @@
-import { Distribution, resultMap } from "../../src/js/index";
+import { Distribution, resultMap, defaultBindings } from "../../src/js/index";
 import { testRun, testRunPartial } from "./TestHelpers";
 
 function Ok<b>(x: b) {
@@ -70,20 +70,19 @@ describe("Partials", () => {
 
 describe("Parameters", () => {
   test("Can pass parameters into partials and cells", () => {
-    let bindings = testRunPartial(`y = $x + 2`, {}, { x: 1 }); // y = 3
-    let bindings2 = testRunPartial(`z = $x + y * $a`, bindings, { a: 3 }); // z = 1 + 3 * 3 = 10
-    expect(testRun(`z + $x + $a + y`, bindings2)).toEqual({
+    let bindings = testRunPartial(`y = $x + 2`, defaultBindings, { x: 1 });
+    let bindings2 = testRunPartial(`z = y + $a`, bindings, { a: 3 });
+    expect(testRun(`z`, bindings2)).toEqual({
       tag: "number",
       value: 17,
     });
   });
   test("Complicated deep parameters", () => {
     expect(
-      testRun(
-        `$x.y[0][0].w + $x.z + $u.v`,
-        {},
-        { x: { y: [[{ w: 1 }]], z: 2 }, u: { v: 3 } }
-      )
+      testRun(`$x.y[0][0].w + $x.z + $u.v`, defaultBindings, {
+        x: { y: [[{ w: 1 }]], z: 2 },
+        u: { v: 3 },
+      })
     ).toEqual({
       tag: "number",
       value: 6,
